third_party/blink/web_tests/platform/mac-mac-arm11.0/external/wpt/css/cssom-view/MediaQueryList-addListener-removeListener-expected.txt

This is a testharness.js-based test.
PASS listeners are called when <iframe> is resized
FAIL listeners are called correct number of times assert_equals: expected 1 but got 0
FAIL listeners are called in order they were added assert_array_equals: lengths differ, expected array ["1st", "2nd"] length 2, got [] length 0
FAIL listener that was added twice is called only once assert_equals: expected 1 but got 0
FAIL listeners are called in order their MQLs were created assert_array_equals: lengths differ, expected array ["mql1", "mql2"] length 2, got [] length 0
FAIL removing listener from one MQL doesn't remove it from all MQLs assert_equals: expected 1 but got 0
PASS MediaQueryList::removeListener removes added listener
Harness: the test ran to completion.
